data_migrate_to_redshift_cronjob/requirements.txt
boto3==1.35.79
pandas==2.2.3
psycopg2-binary
python-dotenv==1.0.1
sqlalchemy==2.0.36
redshift_connector==2.1.5
